Moving file_structure tests to new file to align with new datapack submodule structure
